Fix MPI for nestml building. Add nest vars

+++ cerebellum/nest_models/build_models.py
@@ -4,6 +4,7 @@ from os.path import exists, isdir, join, realpath
 
 import appdirs
 from pynestml.frontend.pynestml_frontend import generate_target
+from bsb.services import MPI
 
 _cereb_dirs = appdirs.AppDirs("cerebellum")
 _cache_path = _cereb_dirs.user_cache_dir
@@ -23,27 +24,28 @@ def _build_nest_models(
     :param str module_name: Name of the nest module produced as outcome.
     :param bool redo: Flag to force rebuild the nest models.
     """
-    if not redo:
-        import nest
+    if MPI.get_size() == 1 or MPI.get_rank() == 0:
+        if not redo:
+            import nest
 
-        try:
-            nest.ResetKernel()
-            nest.Install(module_name)
-            # unload the module
-            nest.ResetKernel()
-            return
-        except nest.NESTErrors.DynamicModuleManagementError as e:
-            if "loaded already" in str(e):
+            try:
+                nest.ResetKernel()
+                nest.Install(module_name)
+                # unload the module
+                nest.ResetKernel()
                 return
-    if not (exists(model_dir) and isdir(model_dir)):
-        raise OSError("Model directory does not exist: {}".format(model_dir))
-    if exists(build_dir) and isdir(build_dir):
-        shutil.rmtree(build_dir)
-    makedirs(build_dir)
+            except nest.NESTErrors.DynamicModuleManagementError as e:
+                if "loaded already" in str(e):
+                    return
+        if not (exists(model_dir) and isdir(model_dir)):
+            raise OSError("Model directory does not exist: {}".format(model_dir))
+        if exists(build_dir) and isdir(build_dir):
+            shutil.rmtree(build_dir)
+        makedirs(build_dir)
 
-    generate_target(
-        input_path=model_dir, target_platform="NEST", target_path=build_dir, module_name=module_name
-    )
+        generate_target(
+            input_path=model_dir, target_platform="NEST", target_path=build_dir, module_name=module_name
+        )
 
 
 _build_nest_models()
